Validantions on todo controller

+++ Controllers/TodoController.cs
@@ -6,6 +6,7 @@ using TodoApp.Domain;
 using System.Collections.Generic;
 using System.Linq;
 using AutoMapper;
+using TodoApp.Extensions;
 
 namespace TodoApp.Controllers
 {
@@ -25,39 +26,61 @@ namespace TodoApp.Controllers
         [HttpGet]
         public async Task<IActionResult> GetListAsync()
         {
-            var todos = await _context.Todos.AsNoTracking().ToListAsync();
-            return Ok(todos);
+            try
+            {
+                var todos = await _context.Todos.AsNoTracking().ToListAsync();
+                return Ok(todos);
+            } catch { return StatusCode(500, "Internal server error"); }
         }
 
         [HttpPost]
         public async Task<IActionResult> CreateAsync([FromBody] TodoCreateViewModel input)
         {
-            var todo = _mapper.Map<TodoCreateViewModel, Todo>(input);
-            todo.UserId = 1;
-            var created = await _context.Todos.AddAsync(todo);
-            await _context.SaveChangesAsync();
-            return Ok(created.Entity);
+            if (!ModelState.IsValid) return (BadRequest(ModelState.GetErrors()));
+
+            try
+            {
+                var todo = _mapper.Map<TodoCreateViewModel, Todo>(input);
+                todo.UserId = 1;
+                var created = await _context.Todos.AddAsync(todo);
+                await _context.SaveChangesAsync();
+                return Ok(created.Entity);
+            } catch { return StatusCode(500, "Internal server error"); }
         }
 
         [HttpPut]
         public async Task<IActionResult> UpdateAsync([FromBody] TodoUpdateViewModel input)
         {
-            var todo = await _context.Todos.FirstOrDefaultAsync(x => x.Id == input.Id);
-            todo.Description = input.Description;
-            todo.DeadLine = input.DeadLine;
-            _context.Todos.Update(todo);
-            await _context.SaveChangesAsync();
-            return Ok(todo);
+            if (!ModelState.IsValid) return (BadRequest(ModelState.GetErrors()));
+
+            try
+            {
+                var todo = await _context.Todos.FirstOrDefaultAsync(x => x.Id == input.Id);
+                if (todo == null) return NotFound("Todo not found");
+
+                todo.Description = input.Description;
+                todo.DeadLine = input.DeadLine;
+                _context.Todos.Update(todo);
+                await _context.SaveChangesAsync();
+                return Ok(todo);
+            } catch { return StatusCode(500, "Internal server error"); }
+            
         }
 
         [HttpDelete]
         [Route("{id:int}")]
         public async Task<IActionResult> DeleteAsync([FromRoute] int id)
         {
-            var todo = await _context.Todos.FirstOrDefaultAsync(x => x.Id == id);
-            _context.Todos.Remove(todo);
-            await _context.SaveChangesAsync();
-            return Ok(todo);
+            try
+            {
+                var todo = await _context.Todos.FirstOrDefaultAsync(x => x.Id == id);
+                if (todo == null) return NotFound("Todo not found");
+
+                _context.Todos.Remove(todo);
+                await _context.SaveChangesAsync();
+                return Ok(todo);
+            }
+            catch { return StatusCode(500, "Internal server error"); }
         }
     }
 }
